+++ src/yoroi-wallets/cardano/ShelleyWallet.ts
@@ -304,6 +304,12 @@ export class ShelleyWallet extends Wallet implements WalletInterface {
     this.save()
   }
 
+  async resync() {
+    await this.clear()
+    super.resync()
+    this.save()
+  }
+
   // =================== utils =================== //
 
   private getNetworkConfig(): CardanoHaskellShelleyNetwork {
